feat: AbstractVariableSizeContainerEntity
