+++ src/main/kotlin/us/timinc/mc/cobblemon/pasturecollector/blocks/entities/PastureCollectorBlockEntity.kt
@@ -5,25 +5,16 @@ import com.cobblemon.mod.common.api.drop.ItemDropEntry
 import com.cobblemon.mod.common.block.entity.PokemonPastureBlockEntity
 import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerFactory
 import net.minecraft.core.BlockPos
-import net.minecraft.core.Direction
-import net.minecraft.core.HolderLookup
 import net.minecraft.core.registries.Registries
-import net.minecraft.nbt.CompoundTag
 import net.minecraft.network.chat.Component
-import net.minecraft.network.protocol.Packet
-import net.minecraft.network.protocol.game.ClientGamePacketListener
-import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket
 import net.minecraft.resources.ResourceLocation
 import net.minecraft.server.level.ServerLevel
 import net.minecraft.server.level.ServerPlayer
-import net.minecraft.world.ContainerHelper
-import net.minecraft.world.WorldlyContainer
 import net.minecraft.world.entity.player.Inventory
 import net.minecraft.world.entity.player.Player
 import net.minecraft.world.item.ItemStack
 import net.minecraft.world.level.Level
 import net.minecraft.world.level.block.Block
-import net.minecraft.world.level.block.entity.BlockEntity
 import net.minecraft.world.level.block.state.BlockState
 import net.minecraft.world.level.storage.loot.LootParams
 import net.minecraft.world.level.storage.loot.parameters.LootContextParams
@@ -36,10 +27,13 @@ import us.timinc.mc.cobblemon.pasturecollector.droppers.PastureBlockDropper
 import us.timinc.mc.cobblemon.pasturecollector.inventory.PastureCollectorMenu
 import us.timinc.mc.cobblemon.pasturecollector.network.BlockPosPayload
 
-class PastureCollectorBlockEntity(val pos: BlockPos, state: BlockState) :
+class PastureCollectorBlockEntity(override val pos: BlockPos, state: BlockState) :
     ExtendedScreenHandlerFactory<BlockPosPayload>,
-    BlockEntity(PastureCollectorBlockEntities.PASTURECOLLECTOR_BLOCKENTITYTYPE, pos, state),
-    WorldlyContainer {
+    AbstractVariableSizeContainerEntity<PastureCollectorBlockEntity>(
+        PastureCollectorBlockEntities.PASTURECOLLECTOR_BLOCKENTITYTYPE,
+        pos,
+        state
+    ) {
     companion object {
         const val CONTAINER_SIZE: Int = 4
 
@@ -56,7 +50,11 @@ class PastureCollectorBlockEntity(val pos: BlockPos, state: BlockState) :
         }
     }
 
-    val inventory = VariedSlotContainer(
+    override fun getScreenOpeningData(player: ServerPlayer): BlockPosPayload = BlockPosPayload(pos)
+
+    override fun getDisplayName(): Component = TITLE
+
+    override var inventory = VariedSlotContainer(
         size = CONTAINER_SIZE,
         onUpdate = { ->
             level?.let { onUpdate(it, blockState, blockState) }
@@ -65,31 +63,9 @@ class PastureCollectorBlockEntity(val pos: BlockPos, state: BlockState) :
         onClose = { -> }
     )
 
-    override fun getUpdatePacket(): Packet<ClientGamePacketListener> = ClientboundBlockEntityDataPacket.create(this)
-
-    override fun getScreenOpeningData(player: ServerPlayer): BlockPosPayload = BlockPosPayload(pos)
-
-    override fun getDisplayName(): Component = TITLE
-
     override fun createMenu(syncId: Int, inventory: Inventory, player: Player) = PastureCollectorMenu(
         syncId, inventory, this
     )
-
-    override fun loadAdditional(nbt: CompoundTag, provider: HolderLookup.Provider) {
-        super.loadAdditional(nbt, provider)
-        ContainerHelper.loadAllItems(nbt, inventory.items, provider)
-    }
-
-    override fun saveAdditional(nbt: CompoundTag, provider: HolderLookup.Provider) {
-        ContainerHelper.saveAllItems(nbt, inventory.items, provider)
-        super.saveAdditional(nbt, provider)
-    }
-
-    override fun getUpdateTag(provider: HolderLookup.Provider): CompoundTag {
-        val nbt = super.getUpdateTag(provider)
-        saveAdditional(nbt, provider)
-        return nbt
-    }
 
     fun onUpdate(level: Level, oldState: BlockState, newState: BlockState) {
         level.sendBlockUpdated(blockPos, oldState, newState, Block.UPDATE_CLIENTS)
@@ -169,34 +145,5 @@ class PastureCollectorBlockEntity(val pos: BlockPos, state: BlockState) :
             }
         }
         return listOfNearbyPastures
-    }
-
-    override fun canPlaceItemThroughFace(i: Int, itemStack: ItemStack, direction: Direction?): Boolean = false
-
-    override fun canTakeItemThroughFace(i: Int, itemStack: ItemStack, direction: Direction): Boolean = true
-
-    override fun clearContent() {
-        inventory.clearContent()
-    }
-
-    override fun getContainerSize(): Int = inventory.size
-    override fun isEmpty(): Boolean = getItem(0).isEmpty
-
-    override fun getItem(i: Int): ItemStack = inventory.getItem(i)
-
-    override fun removeItem(i: Int, j: Int): ItemStack = inventory.removeItem(i, j)
-
-    override fun removeItemNoUpdate(i: Int): ItemStack = inventory.removeItemNoUpdate(i)
-
-    override fun setItem(i: Int, itemStack: ItemStack) = inventory.setItem(i, itemStack)
-
-    override fun stillValid(player: Player): Boolean = inventory.stillValid(player)
-
-    override fun getSlotsForFace(direction: Direction): IntArray {
-        val result = IntArray(inventory.size)
-        for (i in result.indices) {
-            result[i] = i
-        }
-        return result
     }
 }
